Split mirror options file into file for each class
