+++ src/BaGetter.Core/Configuration/MirrorAuthenticationType.cs
@@ -1,0 +1,9 @@
+﻿namespace BaGetter.Core;
+
+public enum MirrorAuthenticationType
+{
+    None,
+    Basic,
+    Bearer,
+    Custom
+}
