+++ src/BaGetter.Core/Configuration/MirrorOptions.cs
@@ -90,24 +90,3 @@ public class MirrorOptions : IValidatableObject
         }
     }
 }
-
-public class MirrorAuthenticationOptions
-{
-    public MirrorAuthenticationType Type { get; set; } = MirrorAuthenticationType.None;
-
-    public string Username { get; set; }
-
-    public string Password { get; set; }
-
-    public string Token { get; set; }
-
-    public Dictionary<string, string> CustomHeaders { get; set; } = [];
-}
-
-public enum MirrorAuthenticationType
-{
-    None,
-    Basic,
-    Bearer,
-    Custom
-}
